+++ pmd-core/src/main/java/net/sourceforge/pmd/cpd/impl/JavaCCTokenizer.java
@@ -2,7 +2,7 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.cpd.internal;
+package net.sourceforge.pmd.cpd.impl;
 
 import java.io.IOException;
