+++ pmd-go/src/main/java/net/sourceforge/pmd/cpd/GoTokenizer.java
@@ -7,7 +7,7 @@ package net.sourceforge.pmd.cpd;
 import org.antlr.v4.runtime.CharStream;
 import org.antlr.v4.runtime.Lexer;
 
-import net.sourceforge.pmd.cpd.internal.AntlrTokenizer;
+import net.sourceforge.pmd.cpd.impl.AntlrTokenizer;
 import net.sourceforge.pmd.lang.go.ast.GolangLexer;
 
 public class GoTokenizer extends AntlrTokenizer {
