+++ pmd-tsql/src/main/java/net/sourceforge/pmd/lang/tsql/cpd/TSqlTokenizer.java
@@ -7,7 +7,7 @@ package net.sourceforge.pmd.lang.tsql.cpd;
 import org.antlr.v4.runtime.CharStream;
 import org.antlr.v4.runtime.Lexer;
 
-import net.sourceforge.pmd.cpd.internal.AntlrTokenizer;
+import net.sourceforge.pmd.cpd.impl.AntlrTokenizer;
 import net.sourceforge.pmd.lang.tsql.ast.TSqlLexer;
 
 public class TSqlTokenizer extends AntlrTokenizer {
